requirements-test.txt
# Testing Dependencies for CodeRed Legal Tech Integration Tests

# Core testing framework
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
pytest-timeout==2.2.0

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# Mocking and fixtures
responses==0.24.1
faker==20.1.0

# Database testing
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis testing
redis==5.0.1
fakeredis==2.20.1

# HTTP testing
httpx==0.25.2
requests-mock==1.11.0

# Data validation
pydantic==2.5.2
python-dotenv==1.0.0

# Decimal precision for cost calculations
decimal

# JSON handling
jsonschema==4.20.0

# Date/time testing
python-dateutil==2.8.2
freezegun==1.4.0

# Reporting
pytest-html==4.1.1
pytest-json-report==1.5.0

# Performance testing
pytest-benchmark==4.0.0

# Code quality
black==23.12.1
flake8==6.1.0
mypy==1.7.1
